Move all game history related components to separate folder
