+++ frontend/src/App.tsx
@@ -22,7 +22,7 @@ import PassAndPlay from "./pages/protected/PassAndPlay.tsx";
 import TempRoute from "./pages/TempRoute.tsx";
 import SelectBot from "./pages/protected/SelectBot/SelectBot.tsx";
 import PlayBot from "./pages/protected/PlayBot.tsx";
-import GameHistory from "./pages/protected/GameHistory.tsx";
+import GameHistory from "./pages/protected/GameHistory/GameHistory.tsx";
 import ViewGame from "./pages/protected/ViewGame.tsx";
 import ChallengeWebsocketProvider from "./components/wrappers/ChallengeWebsocketProvider.tsx";
