+++ frontend/src/pages/protected/GameHistory/components/CompletedGameInfo.tsx
